@@ -4,13 +4,13 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=3.0.6 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
+  cupertino_icons: any
   #TODO: Step 1 - Add the rFlutter Alert package as a dependency here and use "Packages get".
 
 dev_dependencies:
