@@ -82,9 +82,10 @@ class _QuizPageState extends State<QuizPage> {
         Expanded(
           child: Padding(
             padding: EdgeInsets.all(15.0),
-            child: FlatButton(
-              textColor: Colors.white,
-              color: Colors.green,
+            child: TextButton(
+              style: TextButton.styleFrom(
+                  backgroundColor: Colors.green
+              ),
               child: Text(
                 'True',
                 style: TextStyle(
@@ -102,8 +103,10 @@ class _QuizPageState extends State<QuizPage> {
         Expanded(
           child: Padding(
             padding: EdgeInsets.all(15.0),
-            child: FlatButton(
-              color: Colors.red,
+            child: TextButton(
+              style: TextButton.styleFrom(
+                  backgroundColor: Colors.red
+              ),
               child: Text(
                 'False',
                 style: TextStyle(
